@@ -2,10 +2,13 @@
 
 namespace App\Models;
 
-use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\HasOne;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
+use Illuminate\Support\Str;
 
 class User extends Authenticatable
 {
@@ -14,19 +17,20 @@ class User extends Authenticatable
     /**
      * The attributes that are mass assignable.
      *
-     * @var array
+     * @var list<string>
      */
     protected $fillable = [
         'role_id',
+        'currency_id',
         'name',
         'email',
         'password',
     ];
 
     /**
-     * The attributes that should be hidden for arrays.
+     * The attributes that should be hidden for serialization.
      *
-     * @var array
+     * @var list<string>
      */
     protected $hidden = [
         'password',
@@ -34,71 +38,103 @@ class User extends Authenticatable
     ];
 
     /**
-     * The attributes that should be cast to native types.
+     * Get the attributes that should be cast.
      *
-     * @var array
+     * @return array<string, string>
      */
-    protected $casts = [
-        'email_verified_at' => 'datetime',
-    ];
+    protected function casts(): array
+    {
+        return [
+            'email_verified_at' => 'datetime',
+            'password' => 'hashed',
+        ];
+    }
 
     /**
      * The "booted" method of the model.
      * Create uuid when a new user is to be created
-     *
-     * @return void
      */
-    protected static function booted()
+    protected static function booted(): void
     {
-        static::creating(function ($user) {
+        static::creating(function (User $user): void {
             // Generate unique uuid for a new user.
-            $user->uuid = (string) \Illuminate\Support\Str::uuid();
+            $user->uuid = (string) Str::uuid();
 
-            // Default new user currency to USD.
-            $user->currency_id = \App\Models\Currency::where('code', 'USD')->first()->id;
+            // Default new user currency to USD if not set.
+            if (empty($user->currency_id)) {
+                $usdCurrency = Currency::where('code', 'USD')->first();
+                $user->currency_id = $usdCurrency?->id ?? 1;
+            }
 
-            // Default role to customer.
-            // $user->role_id = \App\Models\Role::where('name', 'customer')->first()->id;
+            // Default role to customer if not set.
+            if (empty($user->role_id)) {
+                $customerRole = Role::where('name', 'customer')->first();
+                $user->role_id = $customerRole?->id ?? 2;
+            }
         });
     }
 
     /**
-     * Get the full name of the authenticated user.
+     * Get the full name of the user.
      */
-    public function getFullNameAttribute()
+    public function getFullNameAttribute(): string
     {
         return ucfirst($this->name);
     }
 
     /**
-     * Get the Role associated with the authenticated user.
+     * Get the Role associated with the user.
      */
-    public function role()
+    public function role(): BelongsTo
     {
-        return $this->hasOne(Role::class, 'id', 'role_id');
+        return $this->belongsTo(Role::class);
     }
 
     /**
-     * Get the latest currency transaction of the authenticated user.
+     * Get the transactions of the user.
      */
-    public function transactions()
+    public function transactions(): HasMany
     {
-        return $this->hasMany(Transaction::class)->take(200)->latest('created_at');
+        return $this->hasMany(Transaction::class)->latest('created_at');
     }
 
     /**
-     * Get the latest currency transaction of the authenticated user.
+     * Get the latest currency balance of the user.
      */
-    public function latestCurrencyBalance()
+    public function latestCurrencyBalance(): HasOne
     {
-        return $this->hasOne(CurrencyBalance::class)->latest('created_at');
+        return $this->hasOne(CurrencyBalance::class)->latestOfMany();
     }
 
     /**
-     * Get the default currecny of the authenticated user.
+     * Get all currency balances of the user.
      */
-    public function currency()
+    public function currencyBalances(): HasMany
     {
-        return $this->hasOne(Currency::class, 'id', 'currency_id');
+        return $this->hasMany(CurrencyBalance::class);
+    }
+
+    /**
+     * Get the default currency of the user.
+     */
+    public function currency(): BelongsTo
+    {
+        return $this->belongsTo(Currency::class);
+    }
+
+    /**
+     * Check if user is an administrator.
+     */
+    public function isAdmin(): bool
+    {
+        return $this->role?->name === 'administrator';
+    }
+
+    /**
+     * Check if user is a customer.
+     */
+    public function isCustomer(): bool
+    {
+        return $this->role?->name === 'customer';
     }
 }
