@@ -2,15 +2,16 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Currency;
+use App\Models\CurrencyBalance;
 use App\Models\Transaction;
-use Illuminate\Http\Request;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\View\View;
 
 class HomeController extends Controller
 {
     /**
      * Create a new controller instance.
-     *
-     * @return void
      */
     public function __construct()
     {
@@ -19,64 +20,72 @@ class HomeController extends Controller
 
     /**
      * Display the authenticated user dashboard.
-     *
-     * @return \Illuminate\Contracts\Support\Renderable
      */
-    public function index()
+    public function index(): View
     {
-        // Get authenticated user information.
         $user = auth()->user();
 
+        $transactions = $user->isAdmin()
+            ? Transaction::with(['user', 'recipient', 'sourceCurrency', 'targetCurrency'])
+                ->latest()
+                ->take(500)
+                ->get()
+            : $user->transactions()
+                ->with(['recipient', 'sourceCurrency', 'targetCurrency'])
+                ->take(200)
+                ->get();
+
         return view('application.index', [
-            'transactions'      => ($user->role_id == 2) ? $user['transactions'] : Transaction::take(500)->latest()->get(),
-            'user'              => $user,
+            'transactions' => $transactions,
+            'user' => $user,
         ]);
     }
 
     /**
-     * Refund dollar account
-     *
-     * @return \Illuminate\Contracts\Support\Renderable
+     * Fund the user's dollar account.
      */
-    public function fundAccount()
+    public function fundAccount(): RedirectResponse
     {
-        // Last currency balance
-        $latestCurrencyBalance = auth()->user()->latestCurrencyBalance;
+        $user = auth()->user();
+        $latestCurrencyBalance = $user->latestCurrencyBalance;
 
-        // If the current dollar currency balance is not equal to zero back
-        if($latestCurrencyBalance->USD != 0){
-            return back();
+        // If no balance exists or USD balance is not zero, redirect back
+        if (!$latestCurrencyBalance) {
+            return back()->with('error', 'Unable to process request. Please contact support.');
         }
 
-        // Default currency ID
-        (int) $currency = \App\Models\Currency::where('code', 'USD')->first()->id;
+        if ((float) $latestCurrencyBalance->USD > 0) {
+            return back()->with('info', 'Your USD account already has a balance.');
+        }
 
-        // Default amount
-        (int) $amount = 1000;
+        $currency = Currency::where('code', 'USD')->firstOrFail();
+        $fundingAmount = 1000.00;
+        $fixedFee = 4.86;
 
-        // Credit a new user with $1000.
+        // Credit the user with $1000
         $transaction = Transaction::create([
-            'user_id'               => auth()->id(),
-            'recipient_id'          => 1,
-            'source_currency_id'    => $currency,
-            'target_currency_id'    => $currency,
-            'amount'                => $amount,
-            'rate'                  => 0.0,
-            'transfer_fee'          => 4.86,
-            'variable_fee'          => 0,
-            'fixed_fee'             => 4.86,
-            'type'                  => Transaction::TYPE['Credit'],
-            'status'                => Transaction::STATUS['Success'],
+            'user_id' => $user->id,
+            'recipient_id' => 1, // Admin/System account
+            'source_currency_id' => $currency->id,
+            'target_currency_id' => $currency->id,
+            'amount' => $fundingAmount,
+            'rate' => 1.0,
+            'transfer_fee' => $fixedFee,
+            'variable_fee' => 0,
+            'fixed_fee' => $fixedFee,
+            'type' => Transaction::TYPE['Credit'],
+            'status' => Transaction::STATUS['Success'],
         ]);
 
-        \App\Models\CurrencyBalance::create([
-            'user_id'           => auth()->id(),
-            'transaction_id'    => $transaction['id'],
-            'USD'               => $amount,
-            'EUR'               => $latestCurrencyBalance->EUR,
-            'NGN'               => $latestCurrencyBalance->NGN,
+        CurrencyBalance::create([
+            'user_id' => $user->id,
+            'transaction_id' => $transaction->id,
+            'USD' => $fundingAmount,
+            'EUR' => $latestCurrencyBalance->EUR,
+            'NGN' => $latestCurrencyBalance->NGN,
         ]);
 
-        return redirect()->route('transaction.create')->with('success', 'Your dollar account has been credited with $1,000');
+        return redirect()->route('transaction.create')
+            ->with('success', 'Your dollar account has been credited with $1,000');
     }
 }
