@@ -2,9 +2,10 @@
 
 namespace App\Models;
 
+use App\Services\Currency as CurrencyService;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use App\Services\Currency as CurrencyService;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class Currency extends Model
 {
@@ -12,18 +13,33 @@ class Currency extends Model
 
     protected $table = 'currencies';
 
-    // public $timestamps = false;
+    protected $fillable = [
+        'name',
+        'code',
+        'symbol',
+    ];
 
-    public function users()
+    /**
+     * Get all users with this currency as default.
+     */
+    public function users(): HasMany
     {
         return $this->hasMany(User::class);
     }
 
     /**
-     * Get the type of currency to be rendered
+     * Get the flag details for this currency.
      */
-    public function flag()
+    public function flag(): object
     {
-        return (new CurrencyService)->flag($this->flag);
+        return (new CurrencyService)->flag((string) $this->id);
+    }
+
+    /**
+     * Get formatted display name.
+     */
+    public function getDisplayNameAttribute(): string
+    {
+        return "{$this->symbol} {$this->name} ({$this->code})";
     }
 }
