@@ -2,36 +2,29 @@
 
 namespace Database\Seeders;
 
-use Illuminate\Database\Seeder;
 use App\Models\Currency;
-use App\Models\User;
+use Illuminate\Database\Seeder;
 
 class CurrencySeeder extends Seeder
 {
     /**
      * Run the database seeds.
-     *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
-        //
-        $currencies = self::getCurrency();
-        foreach ($currencies as $currency) {
-            Currency::create([
-                'name'      => $currency['name'],
-                'symbol'    => $currency['symbol'],
-                'code'      => $currency['code'],
-            ]);
-        }
-    }
-
-    public static function getCurrency()
-    {
-        return [
-            ['code' =>'EUR' , 'name' => 'Euro', 'symbol' => '€' ],
-            ['code' =>'NGN' , 'name' => 'Naira', 'symbol' => '₦' ],
-            ['code' =>'USD' , 'name' => 'US Dollar', 'symbol' => '$' ],
+        $currencies = [
+            ['code' => 'EUR', 'name' => 'Euro', 'symbol' => '€'],
+            ['code' => 'NGN', 'name' => 'Nigerian Naira', 'symbol' => '₦'],
+            ['code' => 'USD', 'name' => 'US Dollar', 'symbol' => '$'],
         ];
+
+        foreach ($currencies as $currency) {
+            Currency::updateOrCreate(
+                ['code' => $currency['code']],
+                $currency
+            );
+        }
+
+        $this->command->info('Currencies seeded: EUR, NGN, USD');
     }
 }
