@@ -8,17 +8,21 @@ class DatabaseSeeder extends Seeder
 {
     /**
      * Seed the application's database.
-     *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         $this->call([
             CurrencySeeder::class,
-            ChargeSeeder::class,
             RoleSeeder::class,
+            ChargeSeeder::class,
             UserSeeder::class,
             TransactionSeeder::class,
         ]);
+
+        $this->command->info('Database seeded successfully!');
+        $this->command->info('');
+        $this->command->info('Demo Credentials:');
+        $this->command->info('Admin: admin@wiseclone.com / password');
+        $this->command->info('User: user@wiseclone.com / password');
     }
 }
